requirements.txt
easyocr==1.7.2
opencv-python==4.11.0.86
pillow==11.2.1
matplotlib==3.10.3
arabic-reshaper==3.0.0
python-bidi==0.6.6
